% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{create_deliverability_test_report}
\alias{create_deliverability_test_report}
\title{Create a new predictive inbox placement test}
\usage{
create_deliverability_test_report(ReportName = NULL, FromEmailAddress,
  Content)
}
\arguments{
\item{ReportName}{A unique name that helps you to identify the predictive inbox placement test when you retrieve the results.}

\item{FromEmailAddress}{[required] The email address that the predictive inbox placement test email was sent from.}

\item{Content}{[required] The HTML body of the message that you sent when you performed the predictive inbox placement test.}
}
\description{
Create a new predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will be handled by various email providers around the world. When you perform a predictive inbox placement test, you provide a sample message that contains the content that you plan to send to your customers. Amazon Pinpoint then sends that message to special email addresses spread across several major email providers. After about 24 hours, the test is complete, and you can use the \code{GetDeliverabilityTestReport} operation to view the results of the test.
}
\section{Accepted Parameters}{
\preformatted{create_deliverability_test_report(
  ReportName = "string",
  FromEmailAddress = "string",
  Content = list(
    Simple = list(
      Subject = list(
        Data = "string",
        Charset = "string"
      ),
      Body = list(
        Text = list(
          Data = "string",
          Charset = "string"
        ),
        Html = list(
          Data = "string",
          Charset = "string"
        )
      )
    ),
    Raw = list(
      Data = raw
    )
  )
)
}
}

