% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{delete_dedicated_ip_pool}
\alias{delete_dedicated_ip_pool}
\title{Delete a dedicated IP pool}
\usage{
delete_dedicated_ip_pool(PoolName)
}
\arguments{
\item{PoolName}{[required] The name of the dedicated IP pool that you want to delete.}
}
\description{
Delete a dedicated IP pool.
}
\section{Accepted Parameters}{
\preformatted{delete_dedicated_ip_pool(
  PoolName = "string"
)
}
}

