% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{delete_email_identity}
\alias{delete_email_identity}
\title{Deletes an email identity that you previously verified for use with Amazon Pinpoint}
\usage{
delete_email_identity(
  EmailIdentity = "string"
)
}
\arguments{
\item{EmailIdentity}{[required] The identity (that is, the email address or domain) that you want to delete from your Amazon Pinpoint account.}
}
\description{
Deletes an email identity that you previously verified for use with Amazon Pinpoint. An identity can be either an email address or a domain name.
}
\examples{

}
