% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{get_account}
\alias{get_account}
\title{Obtain information about the email-sending status and capabilities of your Amazon Pinpoint account in the current AWS Region}
\usage{
get_account()
}
\description{
Obtain information about the email-sending status and capabilities of your Amazon Pinpoint account in the current AWS Region.
}
\section{Accepted Parameters}{
\preformatted{get_account()
}
}

