% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{get_deliverability_dashboard_options}
\alias{get_deliverability_dashboard_options}
\title{Show the status of the Deliverability dashboard}
\usage{
get_deliverability_dashboard_options()
}
\description{
Show the status of the Deliverability dashboard. When the Deliverability dashboard is enabled, you gain access to reputation metrics for the domains that you use to send email using Amazon Pinpoint. You also gain the ability to perform predictive inbox placement tests.
}
\details{
When you use the Deliverability dashboard, you pay a monthly charge of USD\$1,250.00, in addition to any other fees that you accrue by using Amazon Pinpoint. If you enable the Deliverability dashboard after the first day of a calendar month, AWS prorates the monthly charge based on how many days have elapsed in the current calendar month.
}
\examples{

}
