% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{get_deliverability_test_report}
\alias{get_deliverability_test_report}
\title{Retrieve the results of a predictive inbox placement test}
\usage{
get_deliverability_test_report(ReportId)
}
\arguments{
\item{ReportId}{[required] A unique string that identifies the predictive inbox placement test.}
}
\description{
Retrieve the results of a predictive inbox placement test.
}
\section{Accepted Parameters}{
\preformatted{get_deliverability_test_report(
  ReportId = "string"
)
}
}

