% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{get_email_identity}
\alias{get_email_identity}
\title{Provides information about a specific identity associated with your Amazon Pinpoint account, including the identity's verification status, its DKIM authentication status, and its custom Mail-From settings}
\usage{
get_email_identity(
  EmailIdentity = "string"
)
}
\arguments{
\item{EmailIdentity}{[required] The email identity that you want to retrieve details for.}
}
\description{
Provides information about a specific identity associated with your Amazon Pinpoint account, including the identity's verification status, its DKIM authentication status, and its custom Mail-From settings.
}
\examples{

}
