% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_package.R
\docType{package}
\name{paws.pinpointemail-package}
\alias{paws.pinpointemail}
\alias{paws.pinpointemail-package}
\title{paws.pinpointemail: Amazon Pinpoint Email Service}
\description{
This document contains reference information for
    the [Amazon Pinpoint](https://aws.amazon.com/pinpoint) Email API,
    version 1.0. This document is best used in conjunction with the
    [Amazon Pinpoint Developer
    Guide](http://docs.aws.amazon.com/pinpoint/latest/developerguide/welcome.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
