% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{put_account_dedicated_ip_warmup_attributes}
\alias{put_account_dedicated_ip_warmup_attributes}
\title{Enable or disable the automatic warm-up feature for dedicated IP addresses}
\usage{
put_account_dedicated_ip_warmup_attributes(AutoWarmupEnabled = NULL)
}
\arguments{
\item{AutoWarmupEnabled}{Enables or disables the automatic warm-up feature for dedicated IP addresses that are associated with your Amazon Pinpoint account in the current AWS Region. Set to \code{true} to enable the automatic warm-up feature, or set to \code{false} to disable it.}
}
\description{
Enable or disable the automatic warm-up feature for dedicated IP addresses.
}
\section{Accepted Parameters}{
\preformatted{put_account_dedicated_ip_warmup_attributes(
  AutoWarmupEnabled = TRUE|FALSE
)
}
}

