% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{put_account_sending_attributes}
\alias{put_account_sending_attributes}
\title{Enable or disable the ability of your account to send email}
\usage{
put_account_sending_attributes(
  SendingEnabled = TRUE|FALSE
)
}
\arguments{
\item{SendingEnabled}{Enables or disables your account's ability to send email. Set to \code{true} to enable email sending, or set to \code{false} to disable email sending.

If AWS paused your account's ability to send email, you can't use this operation to resume your account's ability to send email.}
}
\description{
Enable or disable the ability of your account to send email.
}
\examples{

}
