% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{put_configuration_set_delivery_options}
\alias{put_configuration_set_delivery_options}
\title{Associate a configuration set with a dedicated IP pool}
\usage{
put_configuration_set_delivery_options(ConfigurationSetName,
  SendingPoolName = NULL)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to associate with a dedicated IP pool.}

\item{SendingPoolName}{The name of the dedicated IP pool that you want to associate with the configuration set.}
}
\description{
Associate a configuration set with a dedicated IP pool. You can use dedicated IP pools to create groups of dedicated IP addresses for sending specific types of email.
}
\section{Accepted Parameters}{
\preformatted{put_configuration_set_delivery_options(
  ConfigurationSetName = "string",
  SendingPoolName = "string"
)
}
}

