% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{put_configuration_set_reputation_options}
\alias{put_configuration_set_reputation_options}
\title{Enable or disable collection of reputation metrics for emails that you send using a particular configuration set in a specific AWS Region}
\usage{
put_configuration_set_reputation_options(
  ConfigurationSetName = "string",
  ReputationMetricsEnabled = TRUE|FALSE
)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to enable or disable reputation metric tracking for.}

\item{ReputationMetricsEnabled}{If \code{true}, tracking of reputation metrics is enabled for the configuration set. If \code{false}, tracking of reputation metrics is disabled for the configuration set.}
}
\description{
Enable or disable collection of reputation metrics for emails that you send using a particular configuration set in a specific AWS Region.
}
\examples{

}
