% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{put_configuration_set_tracking_options}
\alias{put_configuration_set_tracking_options}
\title{Specify a custom domain to use for open and click tracking elements in email that you send using Amazon Pinpoint}
\usage{
put_configuration_set_tracking_options(ConfigurationSetName,
  CustomRedirectDomain = NULL)
}
\arguments{
\item{ConfigurationSetName}{[required] The name of the configuration set that you want to add a custom tracking domain to.}

\item{CustomRedirectDomain}{The domain that you want to use to track open and click events.}
}
\description{
Specify a custom domain to use for open and click tracking elements in email that you send using Amazon Pinpoint.
}
\section{Accepted Parameters}{
\preformatted{put_configuration_set_tracking_options(
  ConfigurationSetName = "string",
  CustomRedirectDomain = "string"
)
}
}

