% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pinpointemail_operations.R
\name{put_deliverability_dashboard_option}
\alias{put_deliverability_dashboard_option}
\title{Enable or disable the Deliverability dashboard}
\usage{
put_deliverability_dashboard_option(
  DashboardEnabled = TRUE|FALSE
)
}
\arguments{
\item{DashboardEnabled}{[required] Indicates whether the Deliverability dashboard is enabled. If the value is \code{true}, then the dashboard is enabled.}
}
\description{
Enable or disable the Deliverability dashboard. When you enable the Deliverability dashboard, you gain access to reputation metrics for the domains that you use to send email using Amazon Pinpoint. You also gain the ability to perform predictive inbox placement tests.
}
\details{
When you use the Deliverability dashboard, you pay a monthly charge of USD\$1,250.00, in addition to any other fees that you accrue by using Amazon Pinpoint. If you enable the Deliverability dashboard after the first day of a calendar month, we prorate the monthly charge based on how many days have elapsed in the current calendar month.
}
\examples{

}
